Changed some funcions that deal with environmental variables
It should be posible to add an environmental variable that has no value
Not tested since the  env/export builtins are not done

+++ include/env.h
@@ -6,7 +6,7 @@
 /*   By: gbudau <gbudau@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/09 17:54:37 by gbudau            #+#    #+#             */
-/*   Updated: 2020/12/06 19:41:36 by gbudau           ###   ########.fr       */
+/*   Updated: 2020/12/20 22:25:43 by gbudau           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,6 +31,6 @@ int		unset_env(t_list **environ, char *env);
 void	add_env_front(t_list **environ, char **env);
 int		is_env_format(char c, int flag);
 int		verify_env(const char *str);
-char	**split_env(const char *str, size_t idx);
+char	**split_env(const char *str, int idx);
 
 #endif
